Clean up top level planning interface,   make it clear what to totally gut with the coming resource modelling and memory ladder.
